+++ Robot_Engine/Compute/UpgradeVersion.cs
@@ -25,16 +25,18 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using BH.oM.Reflection.Attributes;
 
 namespace BH.Engine.Adapters.Robot
 {
-    public static partial class Modify
+    public static partial class Compute
     {
 
         /***************************************************/
         /**** Public Methods                            ****/
         /***************************************************/
 
+        [Deprecated("3.1", "Upgrade version methods deprecated. To be moved over to Versioning.")]
         public static List<oM.Structure.Loads.ContourLoad> UpgradeVersion(this List<oM.Adapters.Robot.ContourLoad> loads)
         {
             List<oM.Structure.Loads.ContourLoad> upgradedLoads = new List<oM.Structure.Loads.ContourLoad>();
@@ -47,6 +49,7 @@ namespace BH.Engine.Adapters.Robot
 
         /***************************************************/
 
+        [Deprecated("3.1", "Upgrade version methods deprecated. To be moved over to Versioning.")]
         public static oM.Structure.Loads.ContourLoad UpgradeVersion(this oM.Adapters.Robot.ContourLoad load)
         {
             return new oM.Structure.Loads.ContourLoad
@@ -62,6 +65,7 @@ namespace BH.Engine.Adapters.Robot
 
         /***************************************************/
 
+        [Deprecated("3.1", "Upgrade version methods deprecated. To be moved over to Versioning.")]
         public static List<oM.Structure.Loads.GeometricalLineLoad> UpgradeVersion(this List<oM.Adapters.Robot.GeometricalLineLoad> loads)
         {
             List<oM.Structure.Loads.GeometricalLineLoad> upgradedLoads = new List<oM.Structure.Loads.GeometricalLineLoad>();
@@ -74,6 +78,7 @@ namespace BH.Engine.Adapters.Robot
 
         /***************************************************/
 
+        [Deprecated("3.1", "Upgrade version methods deprecated. To be moved over to Versioning.")]
         public static oM.Structure.Loads.GeometricalLineLoad UpgradeVersion(this oM.Adapters.Robot.GeometricalLineLoad load)
         {
             return new oM.Structure.Loads.GeometricalLineLoad
