+++ Robot_Engine/Robot_Engine.csproj
@@ -109,11 +109,11 @@
     <Compile Include="Create\RobotConfig.cs" />
     <Compile Include="Create\DatabaseSettings.cs" />
     <Compile Include="Create\FramingElementDesignProperties.cs" />
-    <Compile Include="Modify\Upgrade.cs" />
+    <Compile Include="Compute\UpgradeVersion.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Query\GetStringFromEnum.cs" />
     <Compile Include="Query\RobotGravityConstant.cs" />
-    <Compile Include="Query\ToDictionaryWithDistinctCheck.cs" />
+    <Compile Include="Query\ToDictionaryDistinctCheck.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Robot_oM\Robot_oM.csproj">
@@ -123,6 +123,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Modify\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
