+++ Robot_oM/Enums/SectionDB.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -26,15 +26,6 @@ namespace BH.oM.Adapters.Robot
     /****               Public Enums                ****/
     /***************************************************/
 
-    public enum MaterialDB
-    {
-        American,
-        British,
-        Eurocode        
-    }
-
-    /***************************************************/
-
     public enum SectionDB
     {
         UKST,
@@ -42,47 +33,5 @@ namespace BH.oM.Adapters.Robot
     }
 
     /***************************************************/
-
-    public enum ObjectProperties
-    {
-        Name, 
-        Number
-    }
-
-    /***************************************************/
-
-    public enum BarProperties
-    {
-        FramingElementDesignProperties,
-        StructureObjectType                
-    }
-
-    /***************************************************/
-
-    public enum NodeProperties
-    {
-        CoordinateSystem
-    }
-
-    /***************************************************/
-
-    public enum Commands
-    {
-
-    }
-
-    /***************************************************/
-
-    public enum DesignCode_Steel
-    {
-        Default = 0,
-        BS5950,
-        BS5950_2000,
-        BS_EN_1993_1_2005_NA_2008_A1_2014,
-        ANSI_AISC_360_10   
-    }
-
-    /***************************************************/
-
 }
